server/requirements.txt
asgiref==3.7.2
autopep8==2.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.11.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.4
